Rename script ran
