+++ src/Forms/Shared/Samples/Network Analysis/TraceUtilityNetwork/TraceUtilityNetwork.xaml
@@ -1,5 +1,5 @@
 ﻿<ContentPage
-    x:Class="ArcGISRuntimeXamarin.Samples.TraceSubnetwork.TraceSubnetwork"
+    x:Class="ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork.TraceUtilityNetwork"
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms">
