+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Network Analysis/TraceUtilityNetwork/TraceUtilityNetwork.xaml
@@ -1,15 +1,15 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.WPF.Samples.TraceSubnetwork.TraceSubnetwork"
+    x:Class="ArcGISRuntime.UWP.Samples.TraceUtilityNetwork.TraceUtilityNetwork"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
+    xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
     <Grid>
-        <esri:MapView x:Name="MyMapView" GeoViewTapped="OnGeoViewTapped" />
+        <esriUI:MapView x:Name="MyMapView" GeoViewTapped="OnGeoViewTapped" />
         <Border
             x:Name="MainUI"
+            Width="450"
             HorizontalAlignment="Right"
             VerticalAlignment="Top"
-            IsEnabled="False"
             Style="{StaticResource BorderStyle}">
             <Grid>
                 <Grid.RowDefinitions>
@@ -50,17 +50,19 @@
                         x:Name="TraceTypes"
                         Grid.Column="1"
                         Margin="5"
-                        HorizontalContentAlignment="Stretch" />
+                        HorizontalAlignment="Stretch" />
                 </Grid>
                 <Button
                     Grid.Row="2"
                     Margin="5"
+                    HorizontalAlignment="Stretch"
                     Click="OnReset"
                     Content="Reset" />
                 <Button
                     Grid.Row="2"
                     Grid.Column="1"
                     Margin="5"
+                    HorizontalAlignment="Stretch"
                     Click="OnTrace"
                     Content="Trace" />
                 <TextBlock
@@ -78,7 +80,6 @@
                     Visibility="Collapsed" />
             </Grid>
         </Border>
-
         <Border
             Name="TerminalPicker"
             HorizontalAlignment="Center"
@@ -93,12 +94,15 @@
                 </Grid.RowDefinitions>
                 <TextBlock
                     Grid.Row="0"
-                    Margin="5"
-                    Text="Select the terminal for this junction." />
+                    Margin="0,0,0,5"
+                    HorizontalAlignment="Center"
+                    Text="Choose the terminal for this junction." />
                 <ComboBox
                     x:Name="Picker"
                     Grid.Row="1"
-                    Margin="5">
+                    Margin="0,5,0,5"
+                    HorizontalAlignment="Stretch"
+                    ItemsSource="{Binding}">
                     <ComboBox.ItemTemplate>
                         <DataTemplate>
                             <TextBlock Text="{Binding Name}" />
@@ -107,10 +111,12 @@
                 </ComboBox>
                 <Button
                     Grid.Row="2"
-                    Margin="5"
+                    Margin="0,5,0,0"
+                    HorizontalAlignment="Stretch"
                     Click="OnTerminalSelected"
                     Content="Select" />
             </Grid>
         </Border>
     </Grid>
 </UserControl>
+
