+++ src/WPF/ArcGISRuntime.WPF.Viewer/ArcGISRuntime.WPF.Viewer.csproj
@@ -146,8 +146,8 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Sample Code -->
-    <Compile Include="Samples\Network Analysis\TraceSubnetwork\TraceSubnetwork.xaml.cs">
-      <DependentUpon>TraceSubnetwork.xaml</DependentUpon>
+    <Compile Include="Samples\Network Analysis\TraceUtilityNetwork\TraceUtilityNetwork.xaml.cs">
+      <DependentUpon>TraceUtilityNetwork.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Network Analysis\NavigateRouteRerouting\GpxProvider.cs" />
     <Compile Include="Samples\Network Analysis\NavigateRouteRerouting\NavigateRouteRerouting.xaml.cs">
@@ -731,7 +731,7 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Sample XAML -->
-    <Page Include="Samples\Network Analysis\TraceSubnetwork\TraceSubnetwork.xaml">
+    <Page Include="Samples\Network Analysis\TraceUtilityNetwork\TraceUtilityNetwork.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -1510,7 +1510,7 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Screenshots -->
-    <Content Include="Samples\Network Analysis\TraceSubnetwork\TraceSubnetwork.jpg">
+    <Content Include="Samples\Network Analysis\TraceUtilityNetwork\TraceUtilityNetwork.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Network Analysis\NavigateRouteRerouting\NavigateRouteRerouting.jpg">
