+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Network Analysis/TraceUtilityNetwork/TraceUtilityNetwork.xaml
@@ -1,15 +1,15 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.UWP.Samples.TraceSubnetwork.TraceSubnetwork"
+    x:Class="ArcGISRuntime.WPF.Samples.TraceUtilityNetwork.TraceUtilityNetwork"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
+    xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
     <Grid>
-        <esriUI:MapView x:Name="MyMapView" GeoViewTapped="OnGeoViewTapped" />
+        <esri:MapView x:Name="MyMapView" GeoViewTapped="OnGeoViewTapped" />
         <Border
             x:Name="MainUI"
-            Width="450"
             HorizontalAlignment="Right"
             VerticalAlignment="Top"
+            IsEnabled="False"
             Style="{StaticResource BorderStyle}">
             <Grid>
                 <Grid.RowDefinitions>
@@ -50,19 +50,17 @@
                         x:Name="TraceTypes"
                         Grid.Column="1"
                         Margin="5"
-                        HorizontalAlignment="Stretch" />
+                        HorizontalContentAlignment="Stretch" />
                 </Grid>
                 <Button
                     Grid.Row="2"
                     Margin="5"
-                    HorizontalAlignment="Stretch"
                     Click="OnReset"
                     Content="Reset" />
                 <Button
                     Grid.Row="2"
                     Grid.Column="1"
                     Margin="5"
-                    HorizontalAlignment="Stretch"
                     Click="OnTrace"
                     Content="Trace" />
                 <TextBlock
@@ -80,6 +78,7 @@
                     Visibility="Collapsed" />
             </Grid>
         </Border>
+
         <Border
             Name="TerminalPicker"
             HorizontalAlignment="Center"
@@ -94,15 +93,12 @@
                 </Grid.RowDefinitions>
                 <TextBlock
                     Grid.Row="0"
-                    Margin="0,0,0,5"
-                    HorizontalAlignment="Center"
-                    Text="Choose the terminal for this junction." />
+                    Margin="5"
+                    Text="Select the terminal for this junction." />
                 <ComboBox
                     x:Name="Picker"
                     Grid.Row="1"
-                    Margin="0,5,0,5"
-                    HorizontalAlignment="Stretch"
-                    ItemsSource="{Binding}">
+                    Margin="5">
                     <ComboBox.ItemTemplate>
                         <DataTemplate>
                             <TextBlock Text="{Binding Name}" />
@@ -111,12 +107,10 @@
                 </ComboBox>
                 <Button
                     Grid.Row="2"
-                    Margin="0,5,0,0"
-                    HorizontalAlignment="Stretch"
+                    Margin="5"
                     Click="OnTerminalSelected"
                     Content="Select" />
             </Grid>
         </Border>
     </Grid>
 </UserControl>
-
